+++ src/Database/Beam/MySQL/Connection.hs
@@ -8,14 +8,13 @@
 
 module Database.Beam.MySQL.Connection where
 
-import           Control.Exception.Safe (Exception, MonadCatch, MonadMask,
-                                         MonadThrow, bracket, throw)
+import           Control.Exception.Safe (Exception, Handler (Handler),
+                                         MonadCatch, MonadMask, MonadThrow,
+                                         bracket, catchAny, catches, throw)
 import           Control.Monad (void)
-import           Control.Monad.Except (ExceptT, MonadError, catchError,
-                                       runExceptT, throwError)
 import           Control.Monad.Free.Church (F, iterM)
 import           Control.Monad.IO.Class (MonadIO (..))
-import           Control.Monad.RWS.Strict (RWS, runRWS)
+import           Control.Monad.RWS.Strict (RWST, evalRWST)
 import           Control.Monad.Reader (MonadReader (ask), ReaderT (..), asks,
                                        runReaderT)
 import           Control.Monad.State.Strict (MonadState (get, put), modify)
@@ -28,6 +27,7 @@ import           Data.Kind (Type)
 import           Data.Proxy (Proxy (Proxy))
 import           Data.Scientific (Scientific)
 import           Data.Text (Text, pack)
+import qualified Data.Text.Encoding as Enc
 import           Data.Text.Lazy (toStrict)
 import           Data.Text.Lazy.Encoding (decodeUtf8)
 import           Data.Time (Day, LocalTime, TimeOfDay)
@@ -41,30 +41,29 @@ import           Database.Beam.Backend (BeamBackend (..), BeamSqlBackend,
                                         FromBackendRowF (..),
                                         FromBackendRowM (..), MonadBeam (..))
 import           Database.Beam.Backend.SQL (BeamSqlBackendIsString, SqlNull)
+import           Database.Beam.MySQL.FromField.DecodeError (DecodeError (DecodeError),
+                                                            Lenient (IEEEInfinity, IEEENaN, IEEETooBig, IEEETooSmall, SomeStrict, TextCouldNotParse),
+                                                            Strict (NotValidJSON, TypeMismatch, UnexpectedNull, Won'tFit))
 import           Database.Beam.MySQL.Syntax.Render (RenderError (..),
                                                     RenderErrorType (..),
                                                     renderDelete, renderInsert,
                                                     renderSelect, renderUpdate)
 import           Database.Beam.MySQL.Utils (toSQLTypeName)
 #ifdef LENIENT
-import           Database.Beam.MySQL.FromField (DecodeError (DecodeError),
-                                                FromFieldLenient (fromFieldLenient),
-                                                Lenient (..), Strict (..))
+import           Database.Beam.MySQL.FromField.Lenient (FromField (fromField))
 #else
-import           Database.Beam.MySQL.FromField (DecodeError (DecodeError),
-                                                FromFieldStrict (fromFieldStrict),
-                                                Strict (..))
+import           Database.Beam.MySQL.FromField.Strict (FromField (fromField))
 #endif
 import           Database.Beam.MySQL.Syntax (MySQLSyntax (..))
 import           Database.Beam.Query (HasQBuilder (..), HasSqlEqualityCheck,
                                       HasSqlQuantifiedEqualityCheck)
 import           Database.Beam.Query.SQL92 (buildSql92Query')
-import           Database.MySQL.Base (ColumnDef, FieldType, MySQLConn,
-                                      MySQLValue (..), Query (..), columnType,
+import           Database.MySQL.Base (ColumnDef, MySQLConn, MySQLValue (..),
+                                      Query (..), columnOrigName, columnType,
                                       execute_, queryVector_, skipToEof)
-import           Prelude hiding (map, mapM, read)
+import           Prelude hiding (mapM, read)
 import           System.IO.Streams (InputStream, read)
-import           System.IO.Streams.Combinators (map, mapM)
+import           System.IO.Streams.Combinators (mapM)
 import qualified System.IO.Streams.List as S
 import           Type.Reflection (TyCon, Typeable, tyConName)
 
@@ -79,15 +78,9 @@ instance BeamSqlBackendIsString MySQL String
 -- | @since 1.2.2.0
 instance BeamSqlBackendIsString MySQL Text
 
-#ifdef LENIENT
--- | @since 1.2.2.0
+-- | @since 1.3.0.0
 instance BeamBackend MySQL where
-  type BackendFromField MySQL = FromFieldLenient
-#else
--- | @since 1.2.2.0
-instance BeamBackend MySQL where
-    type BackendFromField MySQL = FromFieldStrict
-#endif
+  type BackendFromField MySQL = FromField
 
 -- | @since 1.2.2.0
 instance BeamSqlBackend MySQL
@@ -331,36 +324,36 @@ instance Exception MySQLStatementError
 
 -- | Represents various forms of failure on column decodes.
 --
--- @since 1.2.3.0
+-- @since 1.3.0.0
 data ColumnDecodeError =
   -- | We received a @NULL@ for a column that isn't @NULLABLE@.
   --
-  -- @since 1.2.2.0
+  -- @since 1.3.0.0
   FoundUnexpectedNull {
     demandedType   :: {-# UNPACK #-} !Text,
     sqlType        :: {-# UNPACK #-} !Text,
     tablesInvolved :: !(HashSet Text),
-    columnIndex    :: {-# UNPACK #-} !Word
+    columnName     :: {-# UNPACK #-} !Text
     } |
   -- | We cannot decode the given SQL type into the demanded Haskell type.
   --
-  -- @since 1.2.2.0
+  -- @since 1.3.0.0
   Can'tDecodeIntoDemanded {
     demandedType   :: {-# UNPACK #-} !Text,
     sqlType        :: {-# UNPACK #-} !Text,
     tablesInvolved :: !(HashSet Text),
-    columnIndex    :: {-# UNPACK #-} !Word,
+    columnName     :: {-# UNPACK #-} !Text,
     value          :: {-# UNPACK #-} !Text
     } |
   -- | The SQL value provided by the database does not fit into the
   -- representation of the demanded Haskell type.
   --
-  -- @since 1.2.2.0
+  -- @since 1.3.0.0
   ValueWon'tFitIntoType {
     demandedType   :: {-# UNPACK #-} !Text,
     sqlType        :: {-# UNPACK #-} !Text,
     tablesInvolved :: !(HashSet Text),
-    columnIndex    :: {-# UNPACK #-} !Word,
+    columnName     :: {-# UNPACK #-} !Text,
     value          :: {-# UNPACK #-} !Text
     } |
   -- | When parsing leniently, we found a NaN. As the demanded type is allowed
@@ -369,12 +362,12 @@ data ColumnDecodeError =
   --
   -- /See also:/ @LENIENT.md@
   --
-  -- @since 1.2.2.0
+  -- @since 1.3.0.0
   LenientUnexpectedNaN {
     demandedType   :: {-# UNPACK #-} !Text,
     sqlType        :: {-# UNPACK #-} !Text,
     tablesInvolved :: !(HashSet Text),
-    columnIndex    :: {-# UNPACK #-} !Word
+    columnName     :: {-# UNPACK #-} !Text
     } |
   -- | When parsing leniently, we found a negative or positive infinity. As the
   -- demanded type is allowed to vary considerably, we cannot handle this case
@@ -382,12 +375,12 @@ data ColumnDecodeError =
   --
   -- /See also:/ @LENIENT.md@
   --
-  -- @since 1.2.2.0
+  -- @since 1.3.0.0
   LenientUnexpectedInfinity {
     demandedType   :: {-# UNPACK #-} !Text,
     sqlType        :: {-# UNPACK #-} !Text,
     tablesInvolved :: !(HashSet Text),
-    columnIndex    :: {-# UNPACK #-} !Word,
+    columnName     :: {-# UNPACK #-} !Text,
     value          :: {-# UNPACK #-} !Text
     } |
   -- | When parsing leniently, we found a value that's too negative to fit into
@@ -395,12 +388,12 @@ data ColumnDecodeError =
   --
   -- /See also:/ @LENIENT.md@
   --
-  -- @since 1.2.2.0
+  -- @since 1.3.0.0
   LenientTooSmallToFit {
     demandedType   :: {-# UNPACK #-} !Text,
     sqlType        :: {-# UNPACK #-} !Text,
     tablesInvolved :: !(HashSet Text),
-    columnIndex    :: {-# UNPACK #-} !Word,
+    columnName     :: {-# UNPACK #-} !Text,
     value          :: {-# UNPACK #-} !Text
     } |
   -- | When parsing leniently, we found a value that's too large (that is, too
@@ -414,7 +407,7 @@ data ColumnDecodeError =
     demandedType   :: {-# UNPACK #-} !Text,
     sqlType        :: {-# UNPACK #-} !Text,
     tablesInvolved :: !(HashSet Text),
-    columnIndex    :: {-# UNPACK #-} !Word,
+    columnName     :: {-# UNPACK #-} !Text,
     value          :: {-# UNPACK #-} !Text
     } |
   -- | When parsing leniently, we could not decode a value of the requested type
@@ -422,31 +415,31 @@ data ColumnDecodeError =
   --
   -- /See also:/ @LENIENT.md@
   --
-  -- @since 1.2.2.0
+  -- @since 1.3.0.0
   LenientTextCouldn'tParse {
     demandedType   :: {-# UNPACK #-} !Text,
     sqlType        :: {-# UNPACK #-} !Text,
     tablesInvolved :: !(HashSet Text),
-    columnIndex    :: {-# UNPACK #-} !Word,
+    columnName     :: {-# UNPACK #-} !Text,
     value          :: {-# UNPACK #-} !Text
     } |
   -- | There were not enough fields returned by the database to satisfy the
   -- requested type.
   --
-  -- @since 1.2.2.0
+  -- @since 1.3.0.0
   DemandedTooManyFields {
-    fieldsAvailable :: {-# UNPACK #-} !Word,
-    fieldsDemanded  :: {-# UNPACK #-} !Word
+    fieldsAvailable :: {-# UNPACK #-} !Int,
+    fieldsDemanded  :: {-# UNPACK #-} !Int
     } |
   -- | When using the 'ViaJson' newtype wrapper to aid parsing, something went
   -- wrong with parsing the JSON representation of the desired value.
   --
-  -- @since 1.2.2.0
+  -- @since 1.3.0.0
   JSONError {
     demandedType   :: {-# UNPACK #-} !Text,
     sqlType        :: {-# UNPACK #-} !Text,
     tablesInvolved :: !(HashSet Text),
-    columnIndex    :: {-# UNPACK #-} !Word,
+    columnName     :: {-# UNPACK #-} !Text,
     value          :: {-# UNPACK #-} !Text
     }
   deriving stock (
@@ -480,30 +473,30 @@ newtype MySQLM (a :: Type) = MySQLM (ReaderT MySQLMEnv IO a)
 instance MonadFail MySQLM where
   fail err = error ("Internal error with: " <> err)
 
--- | @since 1.2.2.0
+-- | @since 1.3.0.0
 instance MonadBeam MySQL MySQLM where
-  {-# INLINABLE runNoReturn #-}
+  {-# INLINEABLE runNoReturn #-}
   runNoReturn sql = do
     (stmt, conn, _) <- processAndLog sql
     void . liftIO . execute_ conn $ stmt
-  {-# INLINABLE runReturningOne #-}
+  {-# INLINEABLE runReturningOne #-}
   runReturningOne sql = do
     (stmt, conn, tables) <- processAndLog sql
     bracket (acquireStream conn stmt)
             drainStream
             (processOneResult tables)
-  {-# INLINABLE runReturningList #-}
+  {-# INLINEABLE runReturningList #-}
   runReturningList sql = do
     (stmt, conn, tables) <- processAndLog sql
     bracket (acquireStream conn stmt)
             drainStream
-            (\stream -> liftIO (S.toList =<< parseRowsIO tables stream))
-  {-# INLINABLE runReturningMany #-}
+            (\(env, stream) -> liftIO $ S.toList =<< parseRowsIO env tables stream)
+  {-# INLINEABLE runReturningMany #-}
   runReturningMany sql callback = do
     (stmt, conn, tables) <- processAndLog sql
     bracket (acquireStream conn stmt)
             drainStream
-            (\stream -> callback $ liftIO (read =<< parseRowsIO tables stream))
+            (\(env, stream) -> callback $ liftIO (read =<< parseRowsIO env tables stream))
 
 -- | Run a series of MySQL operations without debugging.
 --
@@ -548,182 +541,169 @@ processAndLog sql = do
       pure (stmt, conn, tables)
 
 acquireStream :: (MonadIO m) =>
-  MySQLConn -> Query -> m (InputStream (Vector (FieldType, MySQLValue)))
-acquireStream conn q = liftIO (queryVector_ conn q >>= go)
-  where
-    go :: (Vector ColumnDef, InputStream (Vector MySQLValue)) ->
-          IO (InputStream (Vector (FieldType, MySQLValue)))
-    go (v, stream) = map (V.zip (V.map columnType v)) stream
+  MySQLConn -> Query -> m (Vector ColumnDef, InputStream (Vector MySQLValue))
+acquireStream conn = liftIO . queryVector_ conn
 
-drainStream :: (MonadIO m) => InputStream a -> m ()
-drainStream = liftIO . skipToEof
+drainStream :: (MonadIO m) =>
+  (Vector ColumnDef, InputStream (Vector MySQLValue)) -> m ()
+drainStream (_, stream) = liftIO . skipToEof $ stream
 
 processOneResult :: forall (a :: Type) .
   (FromBackendRow MySQL a) =>
-  HashSet Text -> InputStream (Vector (FieldType, MySQLValue)) -> MySQLM (Maybe a)
-processOneResult tables stream = do
+  HashSet Text ->
+  (Vector ColumnDef, InputStream (Vector MySQLValue)) ->
+  MySQLM (Maybe a)
+processOneResult tables (env, stream) = do
   mRes <- liftIO . read $ stream
   case mRes of
     Nothing  -> pure Nothing
-    Just res ->
-      case runDecode (iterM (decodeFromRow needed) churched) res tables of
-        Left err -> throw err
-        Right x  -> pure . Just $ x
+    Just res -> Just <$> runDecode (iterM (decodeFromRow needed) churched) res tables env
   where
     churched :: F (FromBackendRowF MySQL) a
     FromBackendRowM churched = fromBackendRow
     needed :: Int
-    needed = valuesNeeded (Proxy @MySQL) (Proxy @a)
+    needed = valuesNeeded (Proxy @MySQL) $ Proxy @a
 
 parseRowsIO :: forall (a :: Type) .
   (FromBackendRow MySQL a) =>
+  Vector ColumnDef ->
   HashSet Text ->
-  InputStream (Vector (FieldType, MySQLValue)) ->
+  InputStream (Vector MySQLValue) ->
   IO (InputStream a)
-parseRowsIO tables = mapM go
+parseRowsIO env tables = mapM go
   where
-    go :: Vector (FieldType, MySQLValue) -> IO a
-    go v = case runDecode (iterM (decodeFromRow needed) churched) v tables of
-      Left err -> throw err
-      Right x  -> pure x
+    go :: Vector MySQLValue -> IO a
+    go v = runDecode (iterM (decodeFromRow needed) churched) v tables env
     churched :: F (FromBackendRowF MySQL) a
     FromBackendRowM churched = fromBackendRow
     needed :: Int
-    needed = valuesNeeded (Proxy @MySQL) (Proxy @a)
+    needed = valuesNeeded (Proxy @MySQL) $ Proxy @a
 
 -- Decoding from a row is complex enough to warrant its own operators and an
 -- explicit stack.
 
+data DataExhaustedError = DataExhaustedError {
+  available :: {-# UNPACK #-} !Int,
+  demanded  :: {-# UNPACK #-} !Int
+  }
+  deriving stock (Eq, Show)
+
+instance Exception DataExhaustedError
+
+data LenientDecodeError = LenientDecodeError Int (DecodeError Lenient)
+  deriving stock (Show, Eq)
+
+instance Exception LenientDecodeError
+
+data StrictDecodeError = StrictDecodeError Int (DecodeError Strict)
+  deriving stock (Show, Eq)
+
+instance Exception StrictDecodeError
+
+data DecodeEnv = DecodeEnv {
+  columnInfo :: {-# UNPACK #-} !(Vector ColumnDef),
+  actualData :: {-# UNPACK #-} !(Vector MySQLValue)
+  }
+
+newtype Decode (m :: Type -> Type) (a :: Type) =
+  Decode (RWST DecodeEnv () Int m a)
+  deriving (Functor,
+            Applicative,
+            Monad,
+            MonadReader DecodeEnv,
+            MonadState Int,
+            MonadThrow,
+            MonadCatch) via (RWST DecodeEnv () Int m)
+
+runDecode :: forall (m :: Type -> Type) (a :: Type) .
+  (MonadCatch m) =>
+  Decode m a ->
+  Vector MySQLValue ->
+  HashSet Text ->
+  Vector ColumnDef ->
+  m a
+runDecode (Decode comp) v tables env =
+  catches (fst <$> evalRWST comp (DecodeEnv env v) 0)
+          [Handler handleDataExhausted,
+          Handler handleLenient,
+          Handler handleStrict]
+  where
+    handleDataExhausted :: DataExhaustedError -> m a
+    handleDataExhausted err =
+      throw . DemandedTooManyFields err.available $ err.demanded
+    handleLenient :: LenientDecodeError -> m a
+    handleLenient (LenientDecodeError lastIx (DecodeError err typ)) = do
+      let ft = toSQLTypeName . columnType $ env V.! lastIx
+      let tyName = tyConNameText typ
+      let colName' = Enc.decodeUtf8 . columnOrigName $ env V.! lastIx
+      let v' = pack . show $ v V.! lastIx
+      case err of
+        SomeStrict err' -> handleStrict . StrictDecodeError lastIx . DecodeError err' $ typ
+        IEEENaN -> throw . LenientUnexpectedNaN tyName ft tables $ colName'
+        IEEEInfinity -> throw . LenientUnexpectedInfinity tyName ft tables colName' $ v'
+        IEEETooSmall -> throw . LenientTooSmallToFit tyName ft tables colName' $ v'
+        IEEETooBig -> throw . LenientTooBigToFit tyName ft tables colName' $ v'
+        TextCouldNotParse -> throw . LenientTextCouldn'tParse tyName ft tables colName' $ v'
+    handleStrict :: StrictDecodeError -> m a
+    handleStrict (StrictDecodeError lastIx (DecodeError err typ)) = do
+      let ft = toSQLTypeName . columnType $ env V.! lastIx
+      let tyName = tyConNameText typ
+      let colName' = Enc.decodeUtf8 . columnOrigName $ env V.! lastIx
+      let v' = pack . show $ v V.! lastIx
+      throw $ case err of
+        UnexpectedNull -> FoundUnexpectedNull tyName ft tables colName'
+        TypeMismatch   -> Can'tDecodeIntoDemanded tyName ft tables colName' v'
+        Won'tFit       -> ValueWon'tFitIntoType tyName ft tables colName' v'
+        NotValidJSON   -> JSONError tyName ft tables colName' v'
+
+decodeFromRow :: forall (m :: Type -> Type) (a :: Type) .
+  (MonadCatch m) =>
+  Int -> FromBackendRowF MySQL (Decode m a) -> Decode m a
+decodeFromRow needed = \case
+  ParseOneField callback                                     -> do
+    curr <- currentValue
+    case curr of
+      Nothing  -> do
+        ix <- get
+        throw . DataExhaustedError (ix + 1) $ needed
+      Just val -> case fromField val of
+        Left err -> do
+          ix <- get
+          errBuild ix err
+        Right x  -> advanceIndex >> callback x
+  Alt (FromBackendRowM opt1) (FromBackendRowM opt2) callback -> do
+    ix <- get -- save state
+    catchAny (callback =<< iterM (decodeFromRow needed) opt1)
+             (\err -> do
+                failingIx <- get -- save stage in case we blow up again
+                put ix -- restore our state to how it was
+                -- The 'catch-in-catch' here is needed due to the peculiar way
+                -- in which beam parses NULLable columns. Essentially, it first
+                -- tries to grab a non-NULL value, then, if it fails, tries to
+                -- unconditionally grab a NULL.
+                --
+                -- This is encoded as an Alt. Therefore, if we don't want
+                -- strange false positives regarding NULL parses, we have to
+                -- forward the _first_ error we saw.
+                catchAny (callback =<< iterM (decodeFromRow needed) opt2)
+                         (\_ -> put failingIx >> throw err))
+  FailParseWith err                                          ->
+    error $ "Leaked beam internal with:" <> show err
+  where
 #ifdef LENIENT
-newtype Decode (a :: Type) =
-  Decode (ExceptT (Either Int (DecodeError Lenient))
-          (RWS (Vector MySQLValue) () Int) a)
-  deriving newtype (Functor,
-                    Applicative,
-                    Monad,
-                    MonadError (Either Int (DecodeError Lenient)),
-                    MonadReader (Vector MySQLValue),
-                    MonadState Int)
-
-runDecode :: forall (a :: Type) .
-  Decode a ->
-  Vector (FieldType, MySQLValue) ->
-  HashSet Text ->
-  Either ColumnDecodeError a
-runDecode (Decode comp) v tables =
-  case runRWS (runExceptT comp) (V.map snd v) 0 of
-    (Right res, _, _) -> Right res
-    (Left failure, lastIx, _) -> Left $ case failure of
-      Left needed ->
-        DemandedTooManyFields (fromIntegral lastIx + 1) . fromIntegral $ needed
-      Right (DecodeError err typ)   ->
-        let ft = toSQLTypeName . fst $ v V.! lastIx
-            ix' = fromIntegral lastIx
-            v' = pack . show . snd $ v V.! lastIx
-            tyName = tyConNameText typ in
-          case err of
-            SomeStrict err' -> case err' of
-              UnexpectedNull -> FoundUnexpectedNull tyName ft tables ix'
-              TypeMismatch   -> Can'tDecodeIntoDemanded tyName ft tables ix' v'
-              Won'tFit       -> ValueWon'tFitIntoType tyName ft tables ix' v'
-              NotValidJSON   -> JSONError tyName ft tables ix' v'
-            IEEENaN -> LenientUnexpectedNaN tyName ft tables ix'
-            IEEEInfinity -> LenientUnexpectedInfinity tyName ft tables ix' v'
-            IEEETooSmall -> LenientTooSmallToFit tyName ft tables ix' v'
-            IEEETooBig -> LenientTooBigToFit tyName ft tables ix' v'
-            TextCouldNotParse -> LenientTextCouldn'tParse tyName ft tables ix' v'
-
-decodeFromRow :: Int -> FromBackendRowF MySQL (Decode a) -> Decode a
-decodeFromRow needed = \case
-  ParseOneField callback -> do
-    curr <- currentValue
-    case curr of
-      Nothing -> throwError . Left $ needed
-      Just val -> case fromFieldLenient val of
-        Left err -> throwError . Right $ err
-        Right x  -> advanceIndex >> callback x
-  Alt (FromBackendRowM opt1) (FromBackendRowM opt2) callback -> do
-    ix <- get -- save state
-    -- The 'catch-in-catch' here is needed due to the rather peculiar way beam
-    -- parses NULLable columns. Essentially, it first tries to grab a non-NULL
-    -- value, then, if it fails, tries to unconditionally grab a NULL.
-    --
-    -- This is encoded as an Alt. Therefore, if we don't want strange false
-    -- positives regarding NULL parses, we have to forward the _first_ error we
-    -- saw.
-    catchError (callback =<< iterM (decodeFromRow needed) opt1)
-               (\err -> do
-                  failingIx <- get -- save state in case we blew up
-                  put ix -- restore our state to how it was
-                  catchError (callback =<< iterM (decodeFromRow needed) opt2)
-                             (\_ -> put failingIx >> throwError err))
-  FailParseWith err -> error ("Leaked beam internal with: " <> show err)
+    errBuild :: Int -> DecodeError Lenient -> Decode m a
+    errBuild ix = throw . LenientDecodeError ix
 #else
-newtype Decode (a :: Type) =
-  Decode (ExceptT (Either Int (DecodeError Strict))
-          (RWS (Vector MySQLValue) () Int) a)
-  deriving newtype (Functor,
-                    Applicative,
-                    Monad,
-                    MonadError (Either Int (DecodeError Strict)),
-                    MonadReader (Vector MySQLValue),
-                    MonadState Int)
-
-runDecode :: forall (a :: Type) .
-  Decode a ->
-  Vector (FieldType, MySQLValue) ->
-  HashSet Text ->
-  Either ColumnDecodeError a
-runDecode (Decode comp) v tables =
-  case runRWS (runExceptT comp) (V.map snd v) 0 of
-    (Right res, _, _) -> Right res
-    (Left failure, lastIx, _) -> Left $ case failure of
-      Left needed ->
-        DemandedTooManyFields (fromIntegral lastIx + 1) . fromIntegral $ needed
-      Right (DecodeError err typ) ->
-        let ft = toSQLTypeName . fst $ v V.! lastIx
-            ix' = fromIntegral lastIx
-            v' = pack . show . snd $ v V.! lastIx
-            tyName = tyConNameText typ in
-          case err of
-            UnexpectedNull -> FoundUnexpectedNull tyName ft tables ix'
-            TypeMismatch   -> Can'tDecodeIntoDemanded tyName ft tables ix' v'
-            Won'tFit       -> ValueWon'tFitIntoType tyName ft tables ix' v'
-            NotValidJSON   -> JSONError tyName ft tables ix' v'
-
-decodeFromRow :: Int -> FromBackendRowF MySQL (Decode a) -> Decode a
-decodeFromRow needed = \case
-  ParseOneField callback -> do
-    curr <- currentValue
-    case curr of
-      Nothing -> throwError . Left $ needed
-      Just val -> case fromFieldStrict val of
-        Left err -> throwError . Right $ err
-        Right x  -> advanceIndex >> callback x
-  Alt (FromBackendRowM opt1) (FromBackendRowM opt2) callback -> do
-    ix <- get -- save state
-    -- The 'catch-in-catch' here is needed due to the rather peculiar way beam
-    -- parses NULLable columns. Essentially, it first tries to grab a non-NULL
-    -- value, then, if it fails, tries to unconditionally grab a NULL.
-    --
-    -- This is encoded as an Alt. Therefore, if we don't want strange false
-    -- positives regarding NULL parses, we have to forward the _first_ error we
-    -- saw.
-    catchError (callback =<< iterM (decodeFromRow needed) opt1)
-               (\err -> do
-                  failingIx <- get -- save state in case we blew up
-                  put ix -- restore our state to how it was
-                  catchError (callback =<< iterM (decodeFromRow needed) opt2)
-                             (\_ -> put failingIx >> throwError err))
-  FailParseWith err -> error ("Leaked beam internal with: " <> show err)
+    errBuild :: Int -> DecodeError Strict -> Decode m a
+    errBuild ix = throw . StrictDecodeError ix
 #endif
 
-currentValue :: Decode (Maybe MySQLValue)
+currentValue :: (Monad m) => Decode m (Maybe MySQLValue)
 currentValue = do
   ix <- get
-  asks (V.!? ix)
+  asks ((V.!? ix) . (.actualData))
 
-advanceIndex :: Decode ()
+advanceIndex :: (Monad m) => Decode m ()
 advanceIndex = modify succ
 
 tyConNameText :: TyCon -> Text