+++ test/nullable/Main.hs
@@ -7,7 +7,7 @@ import           Data.Functor.Identity (Identity)
 import qualified Data.HashSet as HS
 import           Database.Beam (SqlSelect, all_, runSelectReturningOne, select)
 import           Database.Beam.MySQL (ColumnDecodeError (FoundUnexpectedNull),
-                                      MySQL, columnIndex, runBeamMySQL,
+                                      MySQL, columnName, runBeamMySQL,
                                       tablesInvolved)
 import           Database.MySQL.Base (MySQLConn, ciCharset, ciDatabase, ciHost,
                                       ciPort, ciUser, close, connect,
@@ -43,7 +43,7 @@ tests conn = testGroup "Unexpected NULL" [
       case res of
         Left cde@FoundUnexpectedNull{} -> do
           assertBool "contains table name" (HS.member "nullable" . tablesInvolved $ cde)
-          assertEqual "indicates correct column" 1 (columnIndex cde)
+          assertEqual "indicates correct column" "data" . columnName $ cde
         Left _ -> assertFailure "Did not indicate we got an unexpected NULL."
         Right _  -> assertFailure "Got a result when not expecting any."